moved CONFIG_BPF_JIT_DISABLED blocks to secondary files
Removed certain CONFIG_BPF_JIT_DISABLED from ebpf_core.c and consolidated logic into ebpf_core_config.c

Refactored execution_context_unit_test.ccp and moved redundant defined(CONFIG_BPF_JIT_DISABLED) blocks to a different file

Tests that are no-op when JIT is disabled only run when JIT is disabled

Refactored libbpf_test.ccp and moved redundant defined(CONFIG_BPF_JIT_DISABLED) blocks to a different file

Removed unecessary functions header from execution context unit test header file

Refactored end_to_end.cpp and moved JIT config specific code to a new file.

Moved JIT config specific code from api_test.cpp to api_test_jit.cpp

Fixed issue where defining multiple function bodies in ebpf_core based on JIT config casued compiler issues.

Fixed multiple issues where declaring function headers in JIT_CONFIG blocks was causing missing definitions.

Changed function names that did not follow non-static naming conventions.

Resolved comments and separated non JIT specific code more clearly.

Resolved include errors in api_test.h.

moved CONFIG_BPF_JIT_DISABLED blocks to secondary files

Removed certain CONFIG_BPF_JIT_DISABLED from ebpf_core.c and consolidated logic into ebpf_core_config.c

Removed certain CONFIG_BPF_JIT_DISABLED from ebpf_core.c and consolidated logic into ebpf_core_config.c

Tests that are no-op when JIT is disabled only run when JIT is disabled

Refactored libbpf_test.ccp and moved redundant defined(CONFIG_BPF_JIT_DISABLED) blocks to a different file

Refactored end_to_end.cpp and moved JIT config specific code to a new file.

Moved JIT config specific code from api_test.cpp to api_test_jit.cpp

Fixed issue where defining multiple function bodies in ebpf_core based on JIT config casued compiler issues.

Fixed multiple issues where declaring function headers in JIT_CONFIG blocks was causing missing definitions.

Changed function names that did not follow non-static naming conventions.

Resolved comments and separated non JIT specific code more clearly.

Resolved include errors in api_test.h.
